@@ -24,64 +32,83 @@
     "filesize": "^6.1.0",
     "lodash": "4.17.15",
     "moment": "2.24.0",
-    "mkdirp": "0.5.1",
-    "runscript": "^1.5.0"
+    "mkdirp": "0.5.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.2",
-    "@babel/eslint-parser": "^7.17.0",
-    "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/plugin-proposal-decorators": "^7.6.0",
-    "@babel/plugin-transform-runtime": "^7.7.4",
-    "@babel/preset-env": "^7.6.2",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/register": "^7.6.2",
-    "@babel/runtime": "^7.7.4",
+    "@applint/spec": "^1.2.0",
+    "@types/jest": "^27.4.1",
+    "@types/lodash": "^4.14.181",
+    "@types/mocha": "^9.1.0",
+    "@types/react": "^18.0.5",
+    "@types/react-dom": "^18.0.1",
+    "@typescript-eslint/parser": "^5.19.0",
     "antd": "^4.9.4",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.6",
-    "babel-polyfill": "^6.26.0",
     "classnames": "^2.2.6",
     "concurrently": "^5.3.0",
+    "cross-env": "^7.0.3",
     "css-loader": "^0.28.8",
     "electron": "17",
     "electron-windows": "17",
     "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-egg": "^5.1.1",
+    "eslint-config-erb": "^4.0.0-alpha.0",
     "eslint-config-prettier": "^6.3.0",
-    "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-compat": "^4.0.2",
+    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-plugin-jest": "^26.1.4",
     "eslint-plugin-mocha": "^4.11.0",
     "eslint-plugin-no-only-tests": "^2.6.0",
+    "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.14.3",
+    "eslint-plugin-react-hooks": "^4.4.0",
     "git-contributor": "*",
     "husky": "*",
+    "jest": "^27.5.1",
     "less": "3.11.3",
     "less-loader": "^6.2.0",
     "mocha": "*",
     "monitor.js": "*",
     "nyc": "*",
+    "prettier": "^2.6.2",
     "prop-types": "^15.7.2",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "style-loader": "^1.1.2",
-    "webpack": "^5.69.1",
-    "webpack-cli": "^4.9.0",
-    "webpack-dev-server": "^4.7.4"
-  },
-  "scripts": {
-    "build": "webpack",
-    "dev": "concurrently 'NODE_ENV=development webpack-dev-server' 'electron ./example/index.js --disable-gpu'",
-    "dev:web": "NODE_ENV=development webpack-dev-server",
-    "test": "nyc --reporter=lcov --reporter=text mocha --recursive",
-    "lint": "eslint . --fix",
-    "contributor": "git-contributor"
+    "ts-jest": "^27.1.4",
+    "ts-loader": "^9.2.8",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.3",
+    "wait-port": "^0.2.9",
+    "webpack": "5.69.1",
+    "webpack-cli": "4.9.0",
+    "webpack-dev-server": "4.7.4"
   },
   "husky": {
     "hooks": {
       "pre-commit": "npm run lint"
     }
+  },
+  "jest": {
+    "testEnvironment": "node",
+    "transform": {
+      "\\.(ts|tsx|js|jsx)$": "ts-jest"
+    },
+    "moduleNameMapper": {
+      "\\.(css|less|sass|scss)$": "identity-obj-proxy"
+    },
+    "moduleFileExtensions": [
+      "js",
+      "jsx",
+      "ts",
+      "tsx",
+      "json"
+    ],
+    "moduleDirectories": [
+      "node_modules"
+    ],
+    "testPathIgnorePatterns": [
+      "dist"
+    ]
   },
   "license": "MIT"
 }
