@@ -1,23 +1,23 @@
-import React, { useState } from 'react';
-import ReactDOM from 'react-dom';
+import { useState } from 'react';
+import { createRoot } from 'react-dom/client';
 import classnames from 'classnames';
 import 'antd/dist/antd.css';
 import StatusBoard from '../../src/StatusBoard';
 import PerfBoard from '../../src/PerfBoard';
-import styles from './main.module.less';
+import styles from './index.module.less';
 
 const container = document.querySelector('#app');
 
 const params = new URLSearchParams(location.search);
 const { ipcRenderer, shell } = window.electron;
 
-const Main = () => {
+const App = () => {
   const [display, setDisplay] = useState(true);
   return (
     <>
       <StatusBoard
-        eventDataChannelName={params.get('EVENT_DATA_CHANNEL_NAME')}
-        eventActionChannelName={params.get('EVENT_ACTION_CHANNEL_NAME')}
+        eventDataChannelName={params.get('EVENT_DATA_CHANNEL_NAME') || ''}
+        eventActionChannelName={params.get('EVENT_ACTION_CHANNEL_NAME') || ''}
         ipcRenderer={ipcRenderer}
         shell={shell}
       />
@@ -28,17 +28,11 @@ const Main = () => {
       >
         <PerfBoard />
         {display ? (
-          <div
-            className={styles.close}
-            onClick={() => setDisplay(false)}
-          >
+          <div className={styles.close} onClick={() => setDisplay(false)}>
             x
           </div>
         ) : (
-          <div
-            className={styles.badge}
-            onClick={() => setDisplay(true)}
-          >
+          <div className={styles.badge} onClick={() => setDisplay(true)}>
             📈
           </div>
         )}
@@ -47,4 +41,5 @@ const Main = () => {
   );
 };
 
-ReactDOM.render(<Main />, container);
+const root = createRoot(container!);
+root.render(<App />);
