@@ -1,33 +1,62 @@
-'use strict';
-
 module.exports = {
-  extends: 'eslint-config-egg',
-  parser: 'babel-eslint',
-  parserOptions: {
-    ecmaVersion: 2018,
+  extends: ['erb', 'plugin:import/recommended'],
+  settings: {
+    react: {
+      version: 'latest',
+    },
   },
   rules: {
-    'valid-jsdoc': 0,
-    'no-script-url': 0,
-    'no-multi-spaces': 0,
-    'default-case': 0,
-    'no-case-declarations': 0,
-    'one-var-declaration-per-line': 0,
-    'no-restricted-syntax': 0,
-    'jsdoc/require-param': 0,
-    'jsdoc/check-param-names': 0,
-    'jsdoc/require-param-description': 0,
-    'arrow-parens': 0,
-    'prefer-promise-reject-errors': 0,
-    'no-control-regex': 0,
-    'no-use-before-define': 0,
-    'array-callback-return': 0,
-    'no-bitwise': 0,
-    'no-self-compare': 0,
-    'one-var': 0,
+    // A temporary hack related to IDE not resolving correct package.json
+    'import/no-extraneous-dependencies': 'off',
+    // Since React 17 and typescript 4.1 you can safely disable the rule
+    'react/react-in-jsx-scope': 'off',
+    '@typescript-eslint/no-shadow': 'warn',
+    'import/extensions': [
+      'error',
+      'ignorePackages',
+      {
+        js: 'never',
+        jsx: 'never',
+        ts: 'never',
+        tsx: 'never',
+      },
+    ],
   },
-  globals: {
-    window: true,
-    document: true,
+  parserOptions: {
+    ecmaVersion: 2020,
+    sourceType: 'module',
+    project: './tsconfig.json',
+    tsconfigRootDir: __dirname,
+    createDefaultProgram: true,
   },
+  settings: {
+    'import/resolver': {
+      // See https://github.com/benmosher/eslint-plugin-import/issues/1396#issuecomment-575727774 for line below
+      node: {},
+      webpack: {
+        config: require.resolve('./webpack.config.js'),
+      },
+      typescript: {},
+    },
+    'import/parsers': {
+      '@typescript-eslint/parser': ['.ts', '.tsx'],
+    },
+  },
+  overrides: [
+    {
+      files: ['**/lib/**'],
+      rules: {
+        'react-hooks/rules-of-hooks': 0,
+      },
+    },
+    {
+      files: ['**/src/**'],
+      rules: {
+        'no-mixed-operators': 1,
+        'react/jsx-closing-tag-location': 1,
+        '@typescript-eslint/no-shadow': 1,
+        '@typescript-eslint/indent': 1,
+      },
+    },
+  ],
 };
