@@ -1,22 +1,16 @@
-'use strict';
+import url from 'url';
+import path from 'path';
+import WindowManager from 'electron-windows';
+import { app } from 'electron';
+import waitPort from 'wait-port';
+import { Monitor } from '../lib/monitor';
+import { PerfTracing } from '../lib/perf/tracing';
+import { EVENT_ACTION_CHANNEL_NAME, EVENT_DATA_CHANNEL_NAME } from '../lib/monitor/constants';
 
-const url = require('url');
-const path = require('path');
-const WindowManager = require('electron-windows');
-const { app } = require('electron');
-const { Monitor, PerfTracing } = require('..');
-
-const monitor = new Monitor({
-  interval: 3 * 1000,
-});
-
-const {
-  EVENT_DATA_CHANNEL_NAME,
-  EVENT_ACTION_CHANNEL_NAME,
-} = Monitor;
+const monitor = new Monitor();
 
 const mainUrl = url.format({
-  pathname: path.join(__dirname, 'renderer', 'main.html'),
+  pathname: path.join(__dirname, 'renderer', 'index.html'),
   protocol: 'file:',
   query: {
     EVENT_DATA_CHANNEL_NAME,
@@ -24,7 +18,10 @@ const mainUrl = url.format({
   },
 });
 
-app.on('ready', () => {
+app.on('ready', async () => {
+  // wait for the port 8000 to be ready
+  await waitPort({ host: 'localhost', port: 8080 });
+
   const windowManager = new WindowManager();
   const win = windowManager.create({
     name: 'main',
@@ -39,10 +36,11 @@ app.on('ready', () => {
       },
     },
   });
+
   win.loadURL(mainUrl);
   win.webContents.openDevTools({ mode: 'detach' });
   win.webContents.on('dom-ready', () => {
-    monitor.on(EVENT_DATA_CHANNEL_NAME, (data) => {
+    monitor.on(EVENT_DATA_CHANNEL_NAME, (data: any) => {
       win.webContents.send(EVENT_DATA_CHANNEL_NAME, data);
     });
     monitor.bindEventToWindow(win);
@@ -54,12 +52,12 @@ app.on('ready', () => {
 
   PerfTracing({
     dumpTargetDir: path.join(process.cwd(), '.electrom'),
-    partitionThreshold: 30E3,
+    partitionThreshold: 30e3,
     memoryDumpConfig: {
       triggers: [
         {
           mode: 'light',
-          periodic_interval_ms: 10E3,
+          periodic_interval_ms: 10e3,
         },
       ],
     },
